refactoring the file name to  google compute engine
