+++ gcp.js
@@ -1,4 +1,4 @@
-const googleCompute = require("./compute/google");
+const googleCompute = require("./compute/google-computeEngine");
 const googleStorage = require("./storage/google-compute");
 const googleStorageBucket = require("./storage/google-storage");
 const googleDNS = require("./network/google-dns");
